requirements.txt
PySide6
pyqtgraph
pyserial
firebase-admin
requests
pytest
python-dotenv